+++ configuration/settings.py
@@ -1,7 +1,7 @@
 import datetime
 from dataclasses import dataclass
 
-__all__ = ("DataCollectionRetrySettings", "DataCollectionSettings")
+__all__ = ("DataCollectionRetrySettings", "DataCollectionSettings", "DataStorageSettings")
 
 
 @dataclass(eq=False, repr=True)
@@ -17,3 +17,13 @@ class DataCollectionRetrySettings:
     interval_sec: int
     count: int
     delay_between_days_sec: int
+
+
+@dataclass(eq=False, repr=True)
+class DataStorageSettings:
+    enabled: bool
+    host: str
+    port: int
+    user: str
+    password: str
+    db: str