+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2022-12-18
+### Added
+- PostgreSQL data storage for collected data
+
 ## 2022-12-17
 ### Added
 - Ability to request range of days